day 9

+++ src/Day0X.kt
@@ -1,16 +1,16 @@
 fun main() {
 
-    fun part1(input: List<String>): Int {
+    fun part1(input: List<String>): Long {
         return 0
     }
 
-    fun part2(input: List<String>): Int {
+    fun part2(input: List<String>): Long {
         return 0
     }
 
     val exampleInput = readInput("Day0x_example")
-    checkEquals(part1(exampleInput), 0)
-    checkEquals(part2(exampleInput), 0)
+    checkEquals(part1(exampleInput), 0L)
+    checkEquals(part2(exampleInput), 0L)
 
     val input = readInput("Day0x")
     "Solution part 1: ${part1(input)}".println()
